pyproject.toml
[project]
name = "fastapi-with-aws-sqs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.36.10",
    "fastapi[standard]>=0.115.8",
    "pydantic-settings>=2.7.1",
    "python-dotenv>=1.0.1",
]
